Address code review feedback from Gemini
- Cross-platform kimi-cli device ID paths (macOS/Windows/Linux)
- Remove unused httpClient field from KimiAuth struct
- Use int64 instead of float64 for ExpiresAt timestamps
- Export KimiAPIBaseURL from auth package, remove duplicate constant
- Handle sjson.SetBytes errors instead of ignoring them
- Remove dead RefreshKimiToken function (duplicated by executor.Refresh)

+++ internal/auth/kimi/token.go
@@ -38,7 +38,7 @@ type KimiTokenData struct {
 	// TokenType is the type of token, typically "Bearer".
 	TokenType string `json:"token_type"`
 	// ExpiresAt is the Unix timestamp when the token expires.
-	ExpiresAt float64 `json:"expires_at"`
+	ExpiresAt int64 `json:"expires_at"`
 	// Scope is the OAuth2 scope granted to the token.
 	Scope string `json:"scope"`
 }
